semantic/requirements.txt
antlr4-python3-runtime==4.13.1
pip-chill==1.0.3